==> COURAGE/6_Physical_measures/2_Height.Rmd <==
---
title: "Height"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/heightDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/heightDS.R')
```

# Data process

## Finland

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506_height`|
|-|-|
| **Label** | `height`|
| **Table name** |`Courage_Finland_w1`|
| **Units** |`cm`|
| **Missings** |`997 = Disable to measure`<br/>`998 = Refusal`|
| **Description** |  |



```{r assign_fin, echo=F}
opal.assign.table.tibble(o, 'COURAGE_fin_w1','Courage.Courage_Finland_w1',variables=list('q2506_height'), missings = TRUE)
COURAGE_fin_w1 <- opal.execute(o,'COURAGE_fin_w1')
```

```{r local_fin, echo=F}
vbl <- COURAGE_fin_w1$q2506_height
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506_height")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(COURAGE_fin_w1, aes(q2506_height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`

**R script:**

```{r harmo_fin}
COURAGE_ds_fin_w1 <- tibble(id=COURAGE_fin_w1$id)
COURAGE_ds_fin_w1$height <- car::recode(COURAGE_fin_w1$q2506_height, "NA=999; 997=993; 998=997")
COURAGE_ds_fin_w1$height <- labelled(COURAGE_ds_fin_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_fin, echo=F}
vbl <- COURAGE_ds_fin_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_fin_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")

```

#### Validation
```{r crosstabulation_fin, echo=F}
```







## Poland

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506_height`|
|-|-|
| **Label** | `height`|
| **Table name** |`Courage_Poland_w1`|
| **Units** |`cm`|
| **Missings** |`997 = Disable to measure`<br/>`998 = Refusal`|
| **Description** |  |


```{r assign_pol1, echo=F}
opal.assign.table.tibble(o, 'COURAGE_pol_w1','Courage.Courage_Poland_w1',variables=list('q2506_height'), missings = TRUE)
COURAGE_pol_w1 <- opal.execute(o,'COURAGE_pol_w1')
```

```{r local_pol1, echo=F}
vbl <- COURAGE_pol_w1$q2506_height
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506_height")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(COURAGE_pol_w1, aes(q2506_height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```


#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`
* `10<q2506_height < 20 assumed to be measured in decimetres into q2506_height*10`
* `else, q2506_height<60 into 999`
* `1000 >= q2506_height assumed to be measured in milimetres into q2506_height/10`


**R script:**

```{r harmo_pol1}
COURAGE_ds_pol_w1 <- tibble(id=COURAGE_pol_w1$id)
COURAGE_pol_w1$q2506_height[which(COURAGE_pol_w1$q2506_height<20)] <- COURAGE_pol_w1$q2506_height[which(COURAGE_pol_w1$q2506_height<20)]*10
COURAGE_pol_w1$q2506_height[which(COURAGE_pol_w1$q2506_height>=1000)] <- COURAGE_pol_w1$q2506_height[which(COURAGE_pol_w1$q2506_height>=1000)]/10

COURAGE_ds_pol_w1$height <- car::recode(COURAGE_pol_w1$q2506_height, "NA=999; 997=993; 998=997; 0:59=999")

```

#### Statistics of the new harmonised variable
```{r descript_pol1, echo=F}
vbl <- COURAGE_ds_pol_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")

```

#### Validation
```{r crosstabulation_pol1, echo=F}

```






### Wave 2 

#### Study-specific variable description



| **Name** | `q2506_height_s1`|
|-|-|
| **Label** | `height`|
| **Table name** |`Courage_Poland_w2`|
| **Units** |`cm`|
| **Missings** |`997 = Disable to measure`<br/>`998 = Refusal`|
| **Description** |  |



```{r assign_pol2, echo=F}
opal.assign.table.tibble(o, 'COURAGE_pol_w2','Courage.Courage_Poland_w2', variables=list('q2506_height_s1'), missings = TRUE)
COURAGE_pol_w2 <- opal.execute(o,'COURAGE_pol_w2')
```

```{r local_pol2, echo=F}
vbl <- COURAGE_pol_w2$q2506_height_s1
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506_height_s1")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(COURAGE_pol_w2, aes(q2506_height_s1)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`

**R script:**

```{r harmo_pol2}
COURAGE_ds_pol_w2 <- tibble(id=COURAGE_pol_w2$id)
COURAGE_ds_pol_w2$height <- car::recode(COURAGE_pol_w2$q2506_height_s1, "NA=999; 997=993; 998=997")
COURAGE_ds_pol_w2$height <- labelled(COURAGE_ds_pol_w2$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_pol2, echo=F}
vbl <- COURAGE_ds_pol_w2$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w2, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")

```

#### Validation
```{r crosstabulation_pol2, echo=F}

```

### Validation across waves


```{r summ pol, echo=F}

# Creation of longitudinal variance frequencies for each population

m.hds <- list(w1=COURAGE_ds_pol_w1,w2=COURAGE_ds_pol_w2)
m.hds[[1]]$height <- car::recode(m.hds[[1]]$height, "miss_values_vector=NA")
m.hds[[2]]$height <- car::recode(m.hds[[2]]$height, "miss_values_vector=NA")
vbl_name <- hd_vbl
# First wave data
dbb <- m.hds[[1]][,c("id",vbl_name)]
# Merge with next waves data
for(ind in 2:length(m.hds)){
  dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
}
names(dbb) <- c("id", names(m.hds))
dbb$var <- apply(dbb[,2:(length(m.hds)+1)],1,var,na.rm=T)
  
knitr::kable(dbb[which(dbb$var>200),], caption = 'Individuals with difference of height across waves higher than 20 cm')
knitr::kable(dbb[which(dbb$w1<100),], caption = 'Individuals with height lower than 100 cm in wave 1')


```

Since there are only four years of difference between waves, we impute with the most standard height the individuals of the previous list of high differences for which this can be decided.

Since the bmi computed for the two individuals with height lower than 100 cm which not appear in the previous set is extremely high (over 100), their height seem to be an error as they coincide with their reported weights, and they have data only in wave 1, we recode their height as missing.

**R script:**

```{r}
  # Corrections

COURAGE_ds_pol_w1$height[COURAGE_ds_pol_w1$id=="616012742"] <- COURAGE_ds_pol_w2$height[COURAGE_ds_pol_w2$id=="616012742"]
COURAGE_ds_pol_w1$height[COURAGE_ds_pol_w1$id=="616301261"] <- COURAGE_ds_pol_w2$height[COURAGE_ds_pol_w2$id=="616301261"]


COURAGE_ds_pol_w1$height[COURAGE_ds_pol_w1$id=="616040871"] <- 999
COURAGE_ds_pol_w1$height[COURAGE_ds_pol_w1$id=="616038232"] <- 999

```


## Spain

### Wave 1 

#### Study-specific variable description



| **Name** | `q2506_height`|
|-|-|
| **Label** | `height`|
| **Table name** |`Courage_Spain_w1`|
| **Units** |`cm`|
| **Missings** |`997 = Disable to measure`<br/>`998 = Refusal`|
| **Description** |  |



```{r assign_spain1, echo=F}
opal.assign.table.tibble(o, 'COURAGE_spain_w1','Courage.Courage_Spain_w1', variables=list('q2506_height'), missings = TRUE)
COURAGE_spain_w1 <- opal.execute(o,'COURAGE_spain_w1')
```

```{r local_spain1, echo=F}
vbl <- COURAGE_spain_w1$q2506_height
vbl_miss <- c(997,998, 889)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506_height")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(COURAGE_spain_w1, aes(q2506_height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```


#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`
There are values 889 that are going to be considered as mistake instead of 998, therefore they will be recoded as 997.


**R script:**

```{r harmo_spain1}
COURAGE_ds_spain_w1 <- tibble(id=COURAGE_spain_w1$id)
COURAGE_ds_spain_w1$height <- car::recode(COURAGE_spain_w1$q2506_height, "NA=999; 997=993; c(998,889)=997")
COURAGE_ds_spain_w1$height <- labelled(COURAGE_ds_spain_w1$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain1, echo=F}
vbl <- COURAGE_ds_spain_w1$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_spain_w1, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")
```

#### Validation
```{r crosstabulation_spain1, echo=F}
```





### Wave 2 

#### Study-specific variable description



| **Name** | `q2506_height_s1`|
|-|-|
| **Label** | `height`|
| **Table name** |`Courage_Spain_w2_1`|
| **Units** |`cm`|
| **Missings** |`997 = Disable to measure`<br/>`998 = Refusal`|
| **Description** |  |



```{r assign_spain2, echo=F}
opal.assign.table.tibble(o, 'COURAGE_spain_w2','Courage.Courage_Spain_w2_1',variables=list('q2506_height_s1'), missings = TRUE)
COURAGE_spain_w2 <- opal.execute(o,'COURAGE_spain_w2')
```

```{r local_spain2, echo=F}
vbl <- COURAGE_spain_w2$q2506_height_s1
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = "q2506_height_s1")
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(COURAGE_spain_w2, aes(q2506_height_s1)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab("Height") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`


**R script:**

```{r harmo_spain2}
COURAGE_ds_spain_w2 <- tibble(id=COURAGE_spain_w2$id)
COURAGE_ds_spain_w2$height <- car::recode(COURAGE_spain_w2$q2506_height_s1, "0:59=999; NA=999; 997=993; 998=997")
COURAGE_ds_spain_w2$height <- labelled(COURAGE_ds_spain_w2$height, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain2, echo=F}
vbl <- COURAGE_ds_spain_w2$height
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_spain_w2, aes(height)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("height") + ylab("Frequency")

```

#### Validation
There were some unusual values that are lower than 100.
```{r crosstabulation_spain2, echo=F}
kable(COURAGE_spain_w2[which(COURAGE_spain_w2$q2506_height_s1<100),],align = 'c')
```


### Validation across waves


```{r summ spain, echo=F}

# Creation of longitudinal variance frequencies for each population

m.hds <- list(w1=COURAGE_ds_spain_w1,w2=COURAGE_ds_spain_w2)
m.hds[[1]]$height <- car::recode(m.hds[[1]]$height, "miss_values_vector=NA")
m.hds[[2]]$height <- car::recode(m.hds[[2]]$height, "miss_values_vector=NA")
vbl_name <- hd_vbl
# First wave data
dbb <- m.hds[[1]][,c("id",vbl_name)]
# Merge with next waves data
for(ind in 2:length(m.hds)){
  dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
}
names(dbb) <- c("id", names(m.hds))
dbb$var <- apply(dbb[,2:(length(m.hds)+1)],1,var,na.rm=T)
  
knitr::kable(dbb[which(dbb$var>200),], caption = 'Individuals with difference of height across waves higher than 20 cm')


```

Since there are only four years of difference between waves, we impute with the most standard height the individuals of the previous list of high differences for which this can be decided.


**R script:**

```{r}
  # Corrections

COURAGE_ds_spain_w2$height[COURAGE_ds_spain_w2$id=="724000032"] <- COURAGE_ds_spain_w1$height[COURAGE_ds_spain_w1$id=="724000032"]
COURAGE_ds_spain_w2$height[COURAGE_ds_spain_w2$id=="724001122"] <- COURAGE_ds_spain_w1$height[COURAGE_ds_spain_w1$id=="724001122"]
COURAGE_ds_spain_w2$height[COURAGE_ds_spain_w2$id=="724002848"] <- COURAGE_ds_spain_w1$height[COURAGE_ds_spain_w1$id=="724002848"]
COURAGE_ds_spain_w2$height[COURAGE_ds_spain_w2$id=="724004316"] <- COURAGE_ds_spain_w1$height[COURAGE_ds_spain_w1$id=="724004316"]


```



## Summary of descriptive statistics of the harmonised variable accross populations and waves

```{r helpfunctions, echo=F}

labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with continuous data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cont_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for continuous data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of column with summary table categories
  t.summ <- summary(m.hds[[1]][2])[1:6]
  # Adding of missing/no-missing values categories
  t.hds <- c(substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1),labels(Continuous_summary(l.hds[[1]][[2]],missing_values = miss_values_vector)$values_table)[[2]])
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    # First, summary values
    t.summ <- summary(m.hds[[i]][2])[1:6]
    # Next, missing/no-missing values
    t.hds <- cbind(t.hds,c(as.numeric(substr(t.summ,regexpr(":", t.summ, fixed=T)+1,nchar(t.summ))),as.numeric(Continuous_summary(l.hds[[i]][[2]],missing_values = miss_values_vector)$values_table[1,])))
  }
  # Add sample size for each wave/population
  t.hds <- rbind(t.hds,c("sample size", sapply(l.hds,function(wave) length(wave[[1]]))))
  # Add wave/population names
  dimnames(t.hds)[[2]] <- c(dimnames(summary(m.hds[[1]][2]))[[2]],lnames)
  return(t.hds)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


courage.cw <- list(fin_w1 = c("fin_w1","Finland w1"), pol_w1 = c("pol_w1","Poland w1"), pol_w2 = c("pol_w2","Poland w2"), spain_w1 = c("spain_w1","Spain w1"), spain_w2 = c("spain_w2","Spain w2"))

# Consider only harmonised waves
l.hds <- lapply(courage.cw, function(wname) if(exists(paste0("COURAGE_ds_",wname[1]))){wname = list(get(paste0("COURAGE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing summaries
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```





# Quality estimation
No comments.



```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
