==> COURAGE/6_Physical_measures/4_Obesity.Rmd <==
---
title: "Obesity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/obesityDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/obesityDS.R')
```


# Data process

## Finland

### Wave 1

#### Study-specific variable description

| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`courage_fin_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`999 = Missing`|
| **Description** |  |

    
```{r assign_fi1, include=FALSE}

load(paste0(datafolder,"fin_w1/bmi",".RData"))
COURAGE_bmi <- bmi
COURAGE_bmi$bmi <- as.numeric(COURAGE_bmi$bmi)
rm(bmi)

```

```{r local_fi1, echo=F}

vbl <- COURAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_fi1}
COURAGE_ds_fin_w1 <- tibble(id=COURAGE_bmi$id)
COURAGE_ds_fin_w1$obesity <- NA
COURAGE_ds_fin_w1$obesity[which(COURAGE_bmi$bmi %in% miss_values_vector)] <- COURAGE_bmi$bmi[which(COURAGE_bmi$bmi %in% miss_values_vector)]
COURAGE_ds_fin_w1$obesity[which(!COURAGE_bmi$bmi %in% miss_values_vector)] <- car::recode(COURAGE_bmi$bmi[which(!COURAGE_bmi$bmi %in% miss_values_vector)], "lo:29.9999999=0; 30:hi=1; NA=999")
```

#### Statistical description of the new harmonised variable
```{r descript_fi1, echo=F}
vari <- COURAGE_ds_fin_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(COURAGE_ds_fin_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_fi1, echo=F}

```




```{r read_fi1, include=FALSE}
rm(COURAGE_bmi)
```


## Poland

### Wave 1

#### Study-specific variable description

| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`courage_pol_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |

    
```{r assign_pol1, include=FALSE}

load(paste0(datafolder,"pol_w1/bmi",".RData"))
COURAGE_bmi <- bmi
COURAGE_bmi$bmi <- as.numeric(COURAGE_bmi$bmi)
rm(bmi)

```

```{r local_pol1, echo=F}

vbl <- COURAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_pol1}
COURAGE_ds_pol_w1 <- tibble(id=COURAGE_bmi$id)
COURAGE_ds_pol_w1$obesity <- NA
COURAGE_ds_pol_w1$obesity[which(COURAGE_bmi$bmi %in% miss_values_vector)] <- COURAGE_bmi$bmi[which(COURAGE_bmi$bmi %in% miss_values_vector)]
COURAGE_ds_pol_w1$obesity[which(!COURAGE_bmi$bmi %in% miss_values_vector)] <- car::recode(COURAGE_bmi$bmi[which(!COURAGE_bmi$bmi %in% miss_values_vector)], "lo:29.9999999=0; 30:hi=1; NA=999")
```

#### Statistical description of the new harmonised variable
```{r descript_pol1, echo=F}
vari <- COURAGE_ds_pol_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(COURAGE_ds_pol_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_pol1, echo=F}

```




```{r read_pol1, include=FALSE}
rm(COURAGE_bmi)
```



### Wave 2

#### Study-specific variable description

| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`courage_pol_w2_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |

    
```{r assign_pol2, include=FALSE}

load(paste0(datafolder,"pol_w2/bmi",".RData"))
COURAGE_bmi <- bmi
COURAGE_bmi$bmi <- as.numeric(COURAGE_bmi$bmi)
rm(bmi)
```

```{r local_pol2, echo=F}

vbl <- COURAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_pol2}
COURAGE_ds_pol_w2 <- tibble(id=COURAGE_bmi$id)
COURAGE_ds_pol_w2$obesity <- NA
COURAGE_ds_pol_w2$obesity[which(COURAGE_bmi$bmi %in% miss_values_vector)] <- COURAGE_bmi$bmi[which(COURAGE_bmi$bmi %in% miss_values_vector)]
COURAGE_ds_pol_w2$obesity[which(!COURAGE_bmi$bmi %in% miss_values_vector)] <- car::recode(COURAGE_bmi$bmi[which(!COURAGE_bmi$bmi %in% miss_values_vector)], "lo:29.9999999=0; 30:hi=1; NA=999")
```

#### Statistical description of the new harmonised variable
```{r descript_pol2, echo=F}
vari <- COURAGE_ds_pol_w2$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(COURAGE_ds_pol_w2, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_pol2, echo=F}

```




```{r read_pol2, include=FALSE}
rm(COURAGE_bmi)
```



## Spain 

### Wave 1

#### Study-specific variable description

| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`courage_spain_w1_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |


    
```{r assign_spain1, include=FALSE}

load(paste0(datafolder,"spain_w1/bmi",".RData"))
COURAGE_bmi <- bmi
COURAGE_bmi$bmi <- as.numeric(COURAGE_bmi$bmi)
rm(bmi)


```

```{r local_spain1, echo=F}

vbl <- COURAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

#### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_spain1}
COURAGE_ds_spain_w1 <- tibble(id=COURAGE_bmi$id)
COURAGE_ds_spain_w1$obesity <- NA
COURAGE_ds_spain_w1$obesity[which(COURAGE_bmi$bmi %in% miss_values_vector)] <- COURAGE_bmi$bmi[which(COURAGE_bmi$bmi %in% miss_values_vector)]
COURAGE_ds_spain_w1$obesity[which(!COURAGE_bmi$bmi %in% miss_values_vector)] <- car::recode(COURAGE_bmi$bmi[which(!COURAGE_bmi$bmi %in% miss_values_vector)], "lo:29.9999999=0; 30:hi=1; NA=999")
```

#### Statistical description of the new harmonised variable
```{r descript_spain1, echo=F}
vari <- COURAGE_ds_spain_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(COURAGE_ds_spain_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab_spain1, echo=F}

```




```{r read_spain1, include=FALSE}
rm(COURAGE_bmi)
```


### Wave 2

#### Study-specific variable description

| **Name** |`bmi`|
|-|-|
| **Label** |`Body Mass Index calculated as weight (kg)/height(m)^2`|
| **Table name** |`courage_spain_w2_hd`|
| **Units** |`kg/m^2`|
| **Missings** |`993 = Disable to measure`<br/>`997 = Refuse`<br/>`999 = Missing`|
| **Description** |  |


```{r assign_spain2, include=FALSE}

load(paste0(datafolder,"spain_w2/bmi",".RData"))
COURAGE_bmi <- bmi
COURAGE_bmi$bmi <- as.numeric(COURAGE_bmi$bmi)
rm(bmi)

```

```{r local_spain2, echo=F}

vbl <- COURAGE_bmi$bmi
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = "bmi")
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_bmi, aes(bmi)) + geom_histogram(stat="bin", binwidth = 0.4, fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab("BMI") + ylab("Frequency")

```

###### Harmonisation algorithm
To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo_spain2}
COURAGE_ds_spain_w2 <- tibble(id=COURAGE_bmi$id)
COURAGE_ds_spain_w2$obesity <- NA
COURAGE_ds_spain_w2$obesity[which(COURAGE_bmi$bmi %in% miss_values_vector)] <- COURAGE_bmi$bmi[which(COURAGE_bmi$bmi %in% miss_values_vector)]
COURAGE_ds_spain_w2$obesity[which(!COURAGE_bmi$bmi %in% miss_values_vector)] <- car::recode(COURAGE_bmi$bmi[which(!COURAGE_bmi$bmi %in% miss_values_vector)], "lo:29.9999999=0; 30:hi=1; NA=999")
```

###### Statistical description of the new harmonised variable
```{r descript_spain2, echo=F}
vari <- COURAGE_ds_spain_w1$obesity
kable(Categorical_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(COURAGE_ds_spain_w1, aes(x=factor(obesity))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

###### Validation
```{r crosstab_spain2, echo=F}

```




```{r read_spain2, include=FALSE}
rm(COURAGE_bmi)
```






## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable by cohort:







```{r helpfunctions, echo=F}


labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with categorical data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cat_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Labelling of categories (for categorical variables, only non-missing values)
    aux_object[[2]] <- labelled(aux_object[[2]], labels = cat_label[1:2])
  # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for categorical data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of columns with categories and labels
  t.hds <- frq(l.hds[[1]][2])[[1]][,c(1,2)] 
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
  }
  # Add sample size for each wave/population
  t.hds[13,] <- c("n", "sample size", sapply(l.hds,function(wave) length(wave[[1]]))
  )
  # Add wave/population names
  names(t.hds)<-c("val", "label",lnames)
  return(t.hds)
  
}

# Creation of trajectories table for each population

trajectories <- function(m.hds,vbl_name){
  
  # First wave data
  dbb <- m.hds[[1]][,c("id",vbl_name)]
  # Merge with next waves data
  for(ind in 2:length(m.hds)){
    dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
  }
  names(dbb) <- c("id", names(m.hds))
  
  # Glue individual data through all waves into trajectories
  v.dbb <- dbb[,2]
  for(ind in 2:length(m.hds)){
    v.dbb <- paste(v.dbb,dbb[,ind+1],sep="")
  }
  # Trajectories and frequencies
  f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
  return(f.dbb)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


courage.cw <- list(fin_w1 = c("fin_w1","Finland w1"), pol_w1 = c("pol_w1","Poland w1"), pol_w2 = c("pol_w2","Poland w2"), spain_w1 = c("spain_w1","Spain w1"), spain_w2 = c("spain_w2","Spain w2"))

# Consider only harmonised waves
l.hds <- lapply(courage.cw, function(wname) if(exists(paste0("COURAGE_ds_",wname[1]))){wname = list(get(paste0("COURAGE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```






# Quality estimation
Once BMI has been calculated using weight (Kg) and Height (m^2),
Obesity was calculated by using the BMI cutoff>=30kg/m2 following the WHO 
classification. Finally, the resulting BMI variable was coded as follows: 
BMI>30kg/m2 =1 (obese); BMI<30kg/m2 =0 (not obese) to generate two categories.




```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
