==> COURAGE/6_Physical_measures/5_Waist.Rmd <==
---
title: "Waist"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/Rmds/setup_albert.r")
```
<!-- ########################################################## --> 


# Description of DataSchema variable

The description of harmonised variable is the following:
  
* Name: `waist`
* Label: `waist circumference of the participant`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not done`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
* Comments: `Antropomeric measure for waist circumference should be measured in centimeters.`

### COURAGE - Finnish - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `q2508_waist`
* Label: `waist`
* Type of variable: `countinous`
* Missings
    + `997 = Disable to measure`
    + `998 = Refusal`

```{r assign_fin}
opal.assign.table.tibble(o, 'COURAGE_fin_w1_waist','Courage.Courage_Finland_w1',
                         variables=list('q2508_waist'), missings = TRUE)
```

```{r local_fin}
COURAGE_fin_w1_waist <- opal.execute(o,'COURAGE_fin_w1_waist')
Continuous_summary(var = COURAGE_fin_w1_waist$q2508_waist, missing_values = c(997,998))
ggplot(COURAGE_fin_w1_waist, aes(q2508_waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`

R code of the ALGORITHM:

```{r harmo_fin}
COURAGE_ds_fin_w1 <- tibble(id=COURAGE_fin_w1_waist$id)
COURAGE_ds_fin_w1$waist <- car::recode(COURAGE_fin_w1_waist$q2508_waist, "NA=999; 997=993; 998=997")
COURAGE_ds_fin_w1$waist <- labelled(COURAGE_ds_fin_w1$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_fin}
Continuous_summary(var = COURAGE_ds_fin_w1$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(COURAGE_ds_fin_w1, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation_fin}
```






### COURAGE - Polish - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `q2508_waist`
* Label: `waist`
* Type of variable: `countinous`
* Missings
    + `997 = Disable to measure`
    + `998 = Refusal`

```{r assign_pol1}
opal.assign.table.tibble(o, 'COURAGE_pol_w1_waist','Courage.Courage_Poland_w1',
                         variables=list('q2508_waist'), missings = TRUE)
```

```{r local_pol1}
COURAGE_pol_w1_waist <- opal.execute(o,'COURAGE_pol_w1_waist')
Continuous_summary(var = COURAGE_pol_w1_waist$q2508_waist, missing_values = c(997,998))
ggplot(COURAGE_pol_w1_waist, aes(q2508_waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

There are values higher than 145 that are not possible. Those values are going to be considered as mistakes and will be recoded as 999 (missing).

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`

R code of the ALGORITHM:

```{r harmo_pol1}
COURAGE_ds_pol_w1 <- tibble(id=COURAGE_pol_w1_waist$id)
COURAGE_ds_pol_w1$waist <- car::recode(COURAGE_pol_w1_waist$q2508_waist, "NA=999; 997=993; 998=997; 600:996=999; 999:hi=999")
COURAGE_ds_pol_w1$waist <- labelled(COURAGE_ds_pol_w1$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_pol1}
Continuous_summary(var = COURAGE_ds_pol_w1$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(COURAGE_ds_pol_w1, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation_pol1}
```





### COURAGE - Polish - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `q2508_waist_s1`
* Label: `waist`
* Type of variable: `countinous`
* Missings
    + `997 = Refusal`
    + `998 = Disable to measure`

```{r assign_pol2}
opal.assign.table.tibble(o, 'COURAGE_pol_w2_waist','Courage.Courage_Poland_w2',
                         variables=list('q2508_waist_s1'), missings = TRUE)
```

```{r local_pol2}
COURAGE_pol_w2_waist <- opal.execute(o,'COURAGE_pol_w2_waist')
Continuous_summary(var = COURAGE_pol_w2_waist$q2508_waist_s1, missing_values = c(997,998))
ggplot(COURAGE_pol_w2_waist, aes(q2508_waist_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`

R code of the ALGORITHM:

```{r harmo_pol2}
COURAGE_ds_pol_w2 <- tibble(id=COURAGE_pol_w2_waist$id)
COURAGE_ds_pol_w2$waist <- car::recode(COURAGE_pol_w2_waist$q2508_waist_s1, "NA=999; 998=993")
COURAGE_ds_pol_w2$waist <- labelled(COURAGE_ds_pol_w2$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_pol2}
Continuous_summary(var = COURAGE_ds_pol_w2$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(COURAGE_ds_pol_w2, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation_pol2}

```







### COURAGE - Spain - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `q2508_waist`
* Label: `waist`
* Type of variable: `countinous`
* Missings
    + `997 = Disable to measure`
    + `998 = Refusal`

```{r assign_spain1}
opal.assign.table.tibble(o, 'COURAGE_spain_w1_waist','Courage.Courage_Spain_w1',
                         variables=list('q2508_waist'), missings = TRUE)
```

```{r local_spain1}
COURAGE_spain_w1_waist <- opal.execute(o,'COURAGE_spain_w1_waist')
Continuous_summary(var = COURAGE_spain_w1_waist$q2508_waist, missing_values = c(997,998))
ggplot(COURAGE_spain_w1_waist, aes(q2508_waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

There are values higher than 889 that are going to be considered as mistakes and will be recoded as 999 (missing).

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 993`
* `998 into 997`


R code of the ALGORITHM:

```{r harmo_spain1}
COURAGE_ds_spain_w1 <- tibble(id=COURAGE_spain_w1_waist$id)
COURAGE_ds_spain_w1$waist <- car::recode(COURAGE_spain_w1_waist$q2508_waist, "NA=999; 997=993; 998=997; 889:996=999")
COURAGE_ds_spain_w1$waist <- labelled(COURAGE_ds_spain_w1$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain1}
Continuous_summary(var = COURAGE_ds_spain_w1$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(COURAGE_ds_spain_w1, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation_spain1}
```



### COURAGE - Spain - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `q2508_waist_s1`
* Label: `waist`
* Type of variable: `countinous`
* Missings
    + `997 = Refusal`
    + `998 = Disable to measure`

```{r assign_spain2}
opal.assign.table.tibble(o, 'COURAGE_spain_w2_waist','Courage.Courage_Spain_w2_1',
                         variables=list('q2508_waist_s1'), missings = TRUE)
```

```{r local_spain2}
COURAGE_spain_w2_waist <- opal.execute(o,'COURAGE_spain_w2_waist')
Continuous_summary(var = COURAGE_spain_w2_waist$q2508_waist_s1, missing_values = c(997,998))
ggplot(COURAGE_spain_w2_waist, aes(q2508_waist_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `997 into 997`
* `998 into 993`

R code of the ALGORITHM:

```{r harmo_spain2}
COURAGE_ds_spain_w2 <- tibble(id=COURAGE_spain_w2_waist$id)
COURAGE_ds_spain_w2$waist <- car::recode(COURAGE_spain_w2_waist$q2508_waist_s1, "NA=999; 998=993")
COURAGE_ds_spain_w2$waist <- labelled(COURAGE_ds_spain_w2$waist, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=994, "Does not answer"=995,  "Not applicable"=996, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain2}
Continuous_summary(var = COURAGE_ds_spain_w2$waist, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(COURAGE_ds_spain_w2, aes(waist)) + geom_histogram(stat="count", fill="steelblue") + xlab("Waist") + ylab("Frequency") + xlim(35,165) 
```

#### Validation
```{r crosstabulation_spain2}
```

#### Quality estimation
No coments.




```{r save, echo=FALSE}
waist <- COURAGE_ds_spain_w1
save(waist, file = "../RData/spain_w1/waist.RData")
rm(waist)

waist <- COURAGE_ds_spain_w2
save(waist, file = "../RData/spain_w2/waist.RData")
rm(waist)

waist <- COURAGE_ds_pol_w1
save(waist, file = "../RData/pol_w1/waist.RData")
rm(waist)

waist <- COURAGE_ds_pol_w2
save(waist, file = "../RData/pol_w2/waist.RData")
rm(waist)

waist <- COURAGE_ds_fin_w1
save(waist, file = "../RData/fin_w1/waist.RData")
rm(waist)
```


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
