==> COURAGE/6_Physical_measures/10_Systolic.Rmd <==
---
title: "Systolic blood pressure"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/Rmds/setup_laia.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `sbp`
* Variable label: `Systolic blood pressure of participant at each wave`
* Variable description: `Systolic blood pressure of participant at each wave`
* Domain: `Physical Measures`
* Value type: `Countinous`
* Comments: `Means of all sistolic blood pressure measures.`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Systolic blood pressure"
ds_label_all <- "Systolic blood pressure of participant at each wave"
```

# Data process

## Finnish

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_syst`
* Label: `Systolic(1)`
* Type of variable: `countinous`

* Name: `q2502_syst`
* Label: `Systolic(2)`
* Type of variable: `countinous`
* Missings
    + `-8.0 = DK`


```{r assign_fin}
opal.assign.table.tibble(o, 'COURAGE_fin_w1_sbp','Courage.Courage_Finland_w1',
                         variables=list('q2501_syst', 'q2502_syst'), missings = TRUE)
COURAGE_fin_w1_sbp <- opal.execute(o,'COURAGE_fin_w1_sbp')
```

```{r local_fin}
vari <- COURAGE_fin_w1_sbp$q2501_syst
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_fin_w1_sbp, aes(q2501_syst)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = COURAGE_fin_w1_sbp$q2502_syst, missing_values = -8)
vari <- COURAGE_fin_w1_sbp$q2502_syst
kable(Continuous_summary(var = vari, missing_values = -8)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = -8)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = -8)$summary, caption = "Summary")
ggplot(COURAGE_fin_w1_sbp, aes(q2502_syst)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`
* `-8 into 998`

R code of the ALGORITHM:

```{r harmo_fin}
COURAGE_fin_w1_sbp$mean_sbp <- apply(X = COURAGE_fin_w1_sbp[,c(2,3)], MARGIN = 1, FUN = mean)

ids_na01 <- which(is.na(COURAGE_fin_w1_sbp$mean_sbp) & is.na(COURAGE_fin_w1_sbp$q2501_syst) & !is.na(COURAGE_fin_w1_sbp$q2502_syst))
ids_na02 <- which(is.na(COURAGE_fin_w1_sbp$mean_sbp) & !is.na(COURAGE_fin_w1_sbp$q2501_syst) & is.na(COURAGE_fin_w1_sbp$q2502_syst))

COURAGE_fin_w1_sbp$mean_sbp[ids_na01] <- COURAGE_fin_w1_sbp$q2502_syst[ids_na01]
COURAGE_fin_w1_sbp$mean_sbp[ids_na02] <- COURAGE_fin_w1_sbp$q2501_syst[ids_na02]

COURAGE_ds_fin_w1 <- tibble(id=COURAGE_fin_w1_sbp$id)
COURAGE_ds_fin_w1$sbp <- car::recode(COURAGE_fin_w1_sbp$mean_sbp, "NA=999; -8=998")
COURAGE_ds_fin_w1$sbp <- labelled(COURAGE_ds_fin_w1$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_fin}
vari <- COURAGE_ds_fin_w1$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_fin_w1, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```


#### Import to OPAL
```{r importOPAL_fin}
#val_labels(COURAGE_ds_fin_w1) <- NULL
#opal.assign.data(o, 'COURAGE_ds_fin_w1', COURAGE_ds_fin_w1)
#opal.symbol_import(o,'COURAGE_ds_fin_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_fin_w1$sbp) <- ds_label_all
sbp <- COURAGE_ds_fin_w1
save(sbp, file = "../RData/fin_w1/sbp.RData")
```






## Polish

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_syst`
* Label: `Systolic(1)`
* Type of variable: `countinous`
* Missings
    + `-8.0 = DK`

* Name: `q2502_syst`
* Label: `Systolic(2)`
* Type of variable: `countinous`


```{r assign_pol1}
opal.assign.table.tibble(o, 'COURAGE_pol_w1_sbp','Courage.Courage_Poland_w1',
                         variables=list('q2501_syst', 'q2502_syst'), missings = TRUE)
COURAGE_pol_w1_sbp <- opal.execute(o,'COURAGE_pol_w1_sbp')
```

```{r local_pol1}
vari <- COURAGE_pol_w1_sbp$q2501_syst
kable(Continuous_summary(var = vari, missing_values = -8)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = -8)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = -8)$summary, caption = "Summary")
ggplot(COURAGE_pol_w1_sbp, aes(q2501_syst)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_pol_w1_sbp$q2502_syst
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) # SI NO HAY ATRIBUTO SHORT-NAME, A MANO
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_pol_w1_sbp, aes(q2502_syst)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

There is a value 0 in the variables 'q2501_syst' and 'q2502_syst' that are going to be considered as mistakes and will be recoded as 999 (missing). 

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`
* `-8 into 998`

R code of the ALGORITHM:

```{r harmo_pol1}
COURAGE_pol_w1_sbp$mean_sbp <- apply(X = COURAGE_pol_w1_sbp[,c(2,3)], MARGIN = 1, FUN = mean)

COURAGE_ds_pol_w1 <- tibble(id=COURAGE_pol_w1_sbp$id)
COURAGE_ds_pol_w1$sbp <- recode(COURAGE_pol_w1_sbp$mean_sbp, "NA=999; 0=999; -8=998")
COURAGE_ds_pol_w1$sbp <- labelled(COURAGE_ds_pol_w1$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_pol1}
vari <-  COURAGE_ds_pol_w1$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w1, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```


#### Import to OPAL
```{r importOPAL_pol1}
#val_labels(COURAGE_ds_pol_w1) <- NULL
#opal.assign.data(o, 'COURAGE_ds_pol_w1', COURAGE_ds_pol_w1)
#opal.symbol_import(o,'COURAGE_ds_pol_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_pol_w1$sbp) <- ds_label_all
sbp <- COURAGE_ds_pol_w1
save(sbp, file = "../RData/pol_w1/sbp.RData")
```






### Wave 2 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_syst_s1`
* Label: `Systolic(1)`
* Type of variable: `countinous`
* Missings
    + `997 = Refusal`
    + `998 = Disable to measure`

* Name: `q2502_syst_s1`
* Label: `Systolic(2)`
* Type of variable: `countinous`

```{r assign_pol2}
opal.assign.table.tibble(o, 'COURAGE_pol_w2_sbp','Courage.Courage_Poland_w2',
                         variables=list('q2501_syst_s1', 'q2502_syst_s1'), missings = TRUE)
COURAGE_pol_w2_sbp <- opal.execute(o,'COURAGE_pol_w2_sbp')
```

```{r local_pol2}
vari <- COURAGE_pol_w2_sbp$q2501_syst_s1
kable(Continuous_summary(var = vari, missing_values = c(997,998))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(997,998))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(997,998))$summary, caption = "Summary")
ggplot(COURAGE_pol_w2_sbp, aes(q2501_syst_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_pol_w2_sbp$q2502_syst_s1
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_pol_w2_sbp, aes(q2502_syst_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep "Refuse" into 997`
* `998 into 993`

R code of the ALGORITHM:

```{r harmo_pol2}
COURAGE_pol_w2_sbp$mean_sbp <- apply(X = COURAGE_pol_w2_sbp[,c(2,3)], MARGIN = 1, FUN = mean)

COURAGE_ds_pol_w2 <- tibble(id=COURAGE_pol_w2_sbp$id)
COURAGE_ds_pol_w2$sbp <- recode(COURAGE_pol_w2_sbp$mean_sbp, "NA=999; 998=993")
COURAGE_ds_pol_w2$sbp <- labelled(COURAGE_ds_pol_w2$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_pol2}
vari <- COURAGE_ds_pol_w2$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w2, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```


#### Import to OPAL
```{r importOPAL_pol2}
#val_labels(COURAGE_ds_pol_w2) <- NULL
#opal.assign.data(o, 'COURAGE_ds_pol_w2', COURAGE_ds_pol_w2)
#opal.symbol_import(o,'COURAGE_ds_pol_w2', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_pol_w2$sbp) <- ds_label_all
sbp <- COURAGE_ds_pol_w2
save(sbp, file = "../RData/pol_w2/sbp.RData")
```







## Spain

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_syst`
* Label: `Systolic(1)`
* Type of variable: `countinous`
* Missings
    + `-8.0 = DK`

* Name: `q2502_syst`
* Label: `Systolic(2)`
* Type of variable: `countinous`
* Missings
    + `-8.0 = DK`

```{r assign_spain1}
opal.assign.table.tibble(o, 'COURAGE_spain_w1_sbp','Courage.Courage_Spain_w1',
                         variables=list('q2501_syst', 'q2502_syst'), missings = TRUE)
COURAGE_spain_w1_sbp <- opal.execute(o,'COURAGE_spain_w1_sbp')
```

```{r local_spain1}
vari <- COURAGE_spain_w1_sbp$q2501_syst
kable(Continuous_summary(var = vari, missing_values = -8)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = -8)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = -8)$summary, caption = "Summary")
ggplot(COURAGE_spain_w1_sbp, aes(q2501_syst)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_spain_w1_sbp$q2502_syst
kable(Continuous_summary(var = vari, missing_values = -8)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = -8)[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = -8)$summary, caption = "Summary")
ggplot(COURAGE_spain_w1_sbp, aes(q2502_syst)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

There are values 889 that are going to be considered as a missing and They are going to be recoded as 999.

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`
* `-8 into 998`

R code of the ALGORITHM:

```{r harmo_spain1}
COURAGE_spain_w1_sbp$mean_sbp <- apply(X = COURAGE_spain_w1_sbp[,c(2,3)], MARGIN = 1, FUN = mean)

ids <- which(COURAGE_spain_w1_sbp$q2501_syst==-8)
COURAGE_spain_w1_sbp$mean_sbp[ids] <- -8

COURAGE_ds_spain_w1 <- tibble(id=COURAGE_spain_w1_sbp$id)
COURAGE_ds_spain_w1$sbp <- recode(COURAGE_spain_w1_sbp$mean_sbp, "NA=999; -8=998")
COURAGE_ds_spain_w1$sbp <- labelled(COURAGE_ds_spain_w1$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain1}
vari <- COURAGE_ds_spain_w1$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_spain_w1, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```


#### Import to OPAL
```{r importOPAL_spain1}
#val_labels(COURAGE_ds_spain_w1) <- NULL
#opal.assign.data(o, 'COURAGE_ds_spain_w1', COURAGE_ds_spain_w1)
#opal.symbol_import(o,'COURAGE_ds_spain_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_spain_w1$sbp) <- ds_label_all
sbp <- COURAGE_ds_spain_w1
save(sbp, file = "../RData/spain_w1/sbp.RData")
```



### Wave 2 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_syst_s1`
* Label: `Systolic(1)`
* Type of variable: `countinous`
* Missings
    + `997 = Refusal`
    + `998 = Disable to measure`

* Name: `q2502_syst_s1`
* Label: `Systolic(2)`
* Type of variable: `countinous`

```{r assign_spain2}
opal.assign.table.tibble(o, 'COURAGE_spain_w2_sbp','Courage.Courage_Spain_w2_1',
                         variables=list('q2501_syst_s1', 'q2502_syst_s1'), missings = TRUE)
COURAGE_spain_w2_sbp <- opal.execute(o,'COURAGE_spain_w2_sbp')
```

```{r local_spain2}
vari <- COURAGE_spain_w2_sbp$q2501_syst_s1
kable(Continuous_summary(var = vari, missing_values = c(997, 998))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(997, 998))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(997, 998))$summary, caption = "Summary")
ggplot(COURAGE_spain_w2_sbp, aes(q2501_syst_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_spain_w2_sbp$q2502_syst_s1 
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_spain_w2_sbp, aes(q2502_syst_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep "Refusal" into 997`
* `998 into 993`


R code of the ALGORITHM:

```{r harmo_spain2}
COURAGE_spain_w2_sbp$mean_sbp <- apply(X = COURAGE_spain_w2_sbp[,c(2,3)], MARGIN = 1, FUN = mean)

ids <- which(COURAGE_spain_w2_sbp$q2501_syst_s1==997 | COURAGE_spain_w2_sbp$q2501_syst_s1==998)
COURAGE_spain_w2_sbp$mean_sbp[ids] <- COURAGE_spain_w2_sbp$q2501_syst_s1[ids]


COURAGE_ds_spain_w2 <- tibble(id=COURAGE_spain_w2_sbp$id)
COURAGE_ds_spain_w2$sbp <- recode(COURAGE_spain_w2_sbp$mean_sbp, "NA=999; 998=993")
COURAGE_ds_spain_w2$sbp <- labelled(COURAGE_ds_spain_w2$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain2}
vari <- COURAGE_ds_spain_w2$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3])
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_spain_w2, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```


#### Import to OPAL
```{r importOPAL_spain2}
#val_labels(COURAGE_ds_spain_w2) <- NULL
#opal.assign.data(o, 'COURAGE_ds_spain_w2', COURAGE_ds_spain_w2)
#opal.symbol_import(o,'COURAGE_ds_spain_w2', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_spain_w2$sbp) <- ds_label_all
sbp <- COURAGE_ds_spain_w2
save(sbp, file = "../RData/spain_w2/sbp.RData")
```



# Quality estimation

Systolic bp was calculated as the mean from the 2 available variables.
