==> COURAGE/6_Physical_measures/11_Diastolic.Rmd <==
---
title: "Diastolic blood pressure"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/Rmds/setup_laia.r")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `dbp`
* Variable label: `Diastolic blood pressure of participantat at each wave`
* Variable description: `Diastolic blood pressure of participantat at each wave`
* Domain: `Physical Measures`
* Value type: `Countinous`
* Comments: `Means of all diastolic blood pressure measures`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Diastolic blood pressure"
ds_label_all <- "Diastolic blood pressure of participantat at each wave"
```


# Data process

## Finnish

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_diast`
* Label: `Diastolic(1)`
* Type of variable: `countinous`

* Name: `q2502_diast`
* Label: `Diastolic(2)`
* Type of variable: `countinous`


```{r assign_fin, echo = FALSE}
opal.assign.table.tibble(o, 'COURAGE_fin_w1_dbp','Courage.Courage_Finland_w1',
                         variables=list('q2501_diast', 'q2502_diast'), missings = TRUE)
COURAGE_fin_w1_dbp <- opal.execute(o,'COURAGE_fin_w1_dbp')
```

```{r local_fin, echo = FALSE}
vari <- COURAGE_fin_w1_dbp$q2501_diast
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_fin_w1_dbp, aes(q2501_diast)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_fin_w1_dbp$q2502_diast
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_fin_w1_dbp, aes(q2502_diast)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`
* `-8 into 998`

**R script:**

```{r harmo_fin}
COURAGE_fin_w1_dbp$mean_dbp <- apply(X = COURAGE_fin_w1_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

ids_na01 <- which(is.na(COURAGE_fin_w1_dbp$q2501_diast) & !is.na(COURAGE_fin_w1_dbp$q2502_diast))
ids_na02 <- which(!is.na(COURAGE_fin_w1_dbp$q2501_diast) & is.na(COURAGE_fin_w1_dbp$q2502_diast))

COURAGE_fin_w1_dbp$mean_dbp[ids_na01] <- COURAGE_fin_w1_dbp$q2502_diast[ids_na01]
COURAGE_fin_w1_dbp$mean_dbp[ids_na02] <- COURAGE_fin_w1_dbp$q2501_diast[ids_na02]

COURAGE_ds_fin_w1 <- tibble(id=COURAGE_fin_w1_dbp$id)
COURAGE_ds_fin_w1$dbp <- recode(COURAGE_fin_w1_dbp$mean_dbp, "NA=999")
COURAGE_ds_fin_w1$dbp <- labelled(COURAGE_ds_fin_w1$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_fin, echo = FALSE}
vari <- COURAGE_ds_fin_w1$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_fin_w1, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation_fin, echo = FALSE}
```

```{r importOPAL_fin, echo = FALSE}
#opal.assign.data(o, 'COURAGE_ds_fin_w1', COURAGE_ds_fin_w1)
#opal.symbol_import(o,'COURAGE_ds_fin_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_fin_w1$dbp) <- ds_label_all
dbp <- COURAGE_ds_fin_w1
save(dbp, file = "../RData/fin_w1/dbp.RData")
```






## Polish

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_diast`
* Label: `Diastolic(1)`
* Type of variable: `countinous`
* Missings
    + `-8.0 = DK`

* Name: `q2502_diast`
* Label: `Diastolic(2)`
* Type of variable: `countinous`


```{r assign_pol1, echo = FALSE}
opal.assign.table.tibble(o, 'COURAGE_pol_w1_dbp','Courage.Courage_Poland_w1',
                         variables=list('q2501_diast', 'q2502_diast'), missings = TRUE)
COURAGE_pol_w1_dbp <- opal.execute(o,'COURAGE_pol_w1_dbp')
```

```{r local_pol1, echo = FALSE}
vari <- COURAGE_pol_w1_dbp$q2501_diast
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_pol_w1_dbp, aes(q2501_diast)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

vari <-  COURAGE_pol_w1_dbp$q2502_diast
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_pol_w1_dbp, aes(q2502_diast)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

There is a value 0 in the variables 'q2501_diast' and 'q2502_diast' that are going to be considered as mistakes and will be recoded as 999 (missing). 

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`

**R script:**

```{r harmo_pol1}
COURAGE_pol_w1_dbp$mean_dbp <- apply(X = COURAGE_pol_w1_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

COURAGE_ds_pol_w1 <- tibble(id=COURAGE_pol_w1_dbp$id)
COURAGE_ds_pol_w1$dbp <- recode(COURAGE_pol_w1_dbp$mean_dbp, "NA=999; 0=999")
COURAGE_ds_pol_w1$dbp <- labelled(COURAGE_ds_pol_w1$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_pol1, echo = FALSE}
vari <- COURAGE_ds_pol_w1$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w1, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
There are a few unusual values form the variable q2502_diast, that are lower than 30. 
```{r crosstabulation_pol1, echo = FALSE}
COURAGE_pol_w1_dbp[which(COURAGE_pol_w1_dbp$q2502_diast<30),]
```

```{r importOPAL_pol1, echo = FALSE}
#opal.assign.data(o, 'COURAGE_ds_pol_w1', COURAGE_ds_pol_w1)
#opal.symbol_import(o,'COURAGE_ds_pol_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_pol_w1$dbp) <- ds_label_all
dbp <- COURAGE_ds_pol_w1
save(dbp, file = "../RData/pol_w1/dbp.RData")
```






###  Wave 2 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_diast_s1`
* Label: `Diastolic(1)`
* Type of variable: `countinous`

* Name: `q2502_diast_s1`
* Label: `Diastolic(2)`
* Type of variable: `countinous`

```{r assign_pol2, echo = FALSE}
opal.assign.table.tibble(o, 'COURAGE_pol_w2_dbp','Courage.Courage_Poland_w2',
                         variables=list('q2501_diast_s1', 'q2502_diast_s1'), missings = TRUE)
COURAGE_pol_w2_dbp <- opal.execute(o,'COURAGE_pol_w2_dbp')
```

```{r local_pol2, echo = FALSE}
vari <- COURAGE_pol_w2_dbp$q2501_diast_s1
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_pol_w2_dbp, aes(q2501_diast_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_pol_w2_dbp$q2502_diast_s1
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_pol_w2_dbp, aes(q2502_diast_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

The variable q2502_diast_s1 has a value 0 that is going to be considered as a mistake and won't take into account to compute the mean of this individual. 

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

**R script:**

```{r harmo_pol2}
COURAGE_pol_w2_dbp$mean_dbp <- apply(X = COURAGE_pol_w2_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

COURAGE_pol_w2_dbp$mean_dbp[which(COURAGE_pol_w2_dbp$q2502_diast_s1==0)] <- COURAGE_pol_w2_dbp$q2501_diast_s1[which(COURAGE_pol_w2_dbp$q2502_diast_s1==0)]

COURAGE_ds_pol_w2 <- tibble(id=COURAGE_pol_w2_dbp$id)
COURAGE_ds_pol_w2$dbp <- recode(COURAGE_pol_w2_dbp$mean_dbp, "NA=999")
COURAGE_ds_pol_w2$dbp <- labelled(COURAGE_ds_pol_w2$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_pol2, echo = FALSE}
vari <- COURAGE_ds_pol_w2$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w2, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation_pol2, echo = FALSE}

```

```{r importOPAL_pol2, echo = FALSE}
#opal.assign.data(o, 'COURAGE_ds_pol_w2', COURAGE_ds_pol_w2)
#opal.symbol_import(o,'COURAGE_ds_pol_w2', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_pol_w2$dbp) <- ds_label_all
dbp <- COURAGE_ds_pol_w2
save(dbp, file = "../RData/pol_w2/dbp.RData")
```







## Spain

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_diast`
* Label: `Diastolic(1)`
* Type of variable: `countinous`

* Name: `q2502_diast`
* Label: `Diastolic(2)`
* Type of variable: `countinous`

```{r assign_spain1, echo = FALSE}
opal.assign.table.tibble(o, 'COURAGE_spain_w1_dbp','Courage.Courage_Spain_w1',
                         variables=list('q2501_diast', 'q2502_diast'), missings = TRUE)
COURAGE_spain_w1_dbp <- opal.execute(o,'COURAGE_spain_w1_dbp')
```

```{r local_spain1, echo = FALSE}
vari <- COURAGE_spain_w1_dbp$q2501_diast
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_spain_w1_dbp, aes(q2501_diast)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_spain_w1_dbp$q2502_diast
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_spain_w1_dbp, aes(q2502_diast)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

There are values 889 that are going to be considered as a missing and They are going to be recoded as 999.

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`
* `-8 into 998`

**R script:**

```{r harmo_spain1}
COURAGE_spain_w1_dbp$mean_dbp <- apply(X = COURAGE_spain_w1_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

ids_na02 <- which(!is.na(COURAGE_spain_w1_dbp$q2501_diast) & is.na(COURAGE_spain_w1_dbp$q2502_diast))

COURAGE_spain_w1_dbp$mean_dbp[ids_na02] <- COURAGE_spain_w1_dbp$q2501_diast[ids_na02]

COURAGE_ds_spain_w1 <- tibble(id=COURAGE_spain_w1_dbp$id)
COURAGE_ds_spain_w1$dbp <- recode(COURAGE_spain_w1_dbp$mean_dbp, "NA=999")
COURAGE_ds_spain_w1$dbp <- labelled(COURAGE_ds_spain_w1$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain1, echo = FALSE}
vari <- COURAGE_ds_spain_w1$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ds_spain_w1, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
There is an unusual values in the variable q2501_diast that is upper than 160.
```{r crosstabulation_spain1, echo = FALSE}
COURAGE_spain_w1_dbp[which(COURAGE_spain_w1_dbp$q2501_diast>160),]
```

```{r importOPAL_spain1, echo = FALSE}
#opal.assign.data(o, 'COURAGE_ds_spain_w1', COURAGE_ds_spain_w1)
#opal.symbol_import(o,'COURAGE_ds_spain_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_spain_w1$dbp) <- ds_label_all
dbp <- COURAGE_ds_spain_w1
save(dbp, file = "../RData/spain_w1/dbp.RData")
```






### Wave 2 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `q2501_diast_s1`
* Label: `Diastolic(1)`
* Type of variable: `countinous`
* Missings
    + `997 = Refusal`
    + `998 = Disable to measure`

* Name: `q2502_diast_s1`
* Label: `Diastolic(2)`
* Type of variable: `countinous`


```{r assign_spain2, echo = FALSE}
opal.assign.table.tibble(o, 'COURAGE_spain_w2_dbp','Courage.Courage_Spain_w2_1',
                         variables=list('q2501_diast_s1', 'q2502_diast_s1'), missings = TRUE)
COURAGE_spain_w2_dbp <- opal.execute(o,'COURAGE_spain_w2_dbp')
```

```{r local_spain2, echo = FALSE}
vari <- COURAGE_spain_w2_dbp$q2501_diast_s1
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_spain_w2_dbp, aes(q2501_diast_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

vari <- COURAGE_spain_w2_dbp$q2502_diast_s1
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_spain_w2_dbp, aes(q2502_diast_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`
* `keep "Refusal" into 997`
* `998 into 993`

**R script:**

```{r harmo_spain2}
COURAGE_spain_w2_dbp$mean_dbp <- apply(X = COURAGE_spain_w2_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

COURAGE_spain_w2_dbp$mean_dbp[COURAGE_spain_w2_dbp$q2501_diast_s1==998] <- 998
COURAGE_spain_w2_dbp$mean_dbp[COURAGE_spain_w2_dbp$q2501_diast_s1==997] <- 997


COURAGE_ds_spain_w2 <- tibble(id=COURAGE_spain_w2_dbp$id)
COURAGE_ds_spain_w2$dbp <- recode(COURAGE_spain_w2_dbp$mean_dbp, "NA=999; 998=993")
COURAGE_ds_spain_w2$dbp <- labelled(COURAGE_ds_spain_w2$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not attempt not done"=996, "Does not answer"=995,  "Not applicable"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript_spain2, echo = FALSE}
Continuous_summary(var = COURAGE_ds_spain_w2$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(COURAGE_ds_spain_w2, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation_spain2, echo = FALSE}


####  Quality estimation
Diastolic  bp was calculated as the mean from the 2 available variables.
```

```{r importOPAL_spain2, echo = FALSE}
#opal.assign.data(o, 'COURAGE_ds_spain_w2', COURAGE_ds_spain_w2)
#opal.symbol_import(o,'COURAGE_ds_spain_w2', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_spain_w2$dbp) <- ds_label_all
dbp <- COURAGE_ds_spain_w2
save(dbp, file = "../RData/spain_w2/dbp.RData")
```
