==> COURAGE/6_Physical_measures/12_ah.Rmd <==
---
title: "Arterial hypertension"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---


```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/Rmds/setup_laia.r")
load("../RData/fin_w1/dbp.RData")
```

# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `ah`
* Variable label: `Arterial hypertension`
* Variable description: `Arterial hypertension. Means of all diastolic blood pressure measures.`
* Domain: `Physical Measures`
* Value type: `Categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | no hypertension
1        | hypertension

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done 
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label_s <- "Systolic blood pressure"
ds_label_d <- "Diastolic blood pressure"
ds_label <- "Arterial hypertension"
ds_label_all <- "Arterial hypertension. Means of all diastolic blood pressure measures."
```

# Data process

## Spain

### Wave 1 

#### Study-specific variable description

The study-specific variables elected to be harmonised is:

* Name: `sbp`
* Label: `Systolic blood pressure of participant at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

* Name: `dbp`
* Label: `Diastolic blood pressure of participantat at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

```{r assign Spain w1, echo=F}
load("../RData/spain_w1/dbp.RData")
COURAGE_ah <- dbp
load("../RData/spain_w1/sbp.RData")
COURAGE_ah$sbp <- sbp$sbp
```

```{r local Spain w1, echo=F}
vari <- COURAGE_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_s)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(x=(sbp))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_s) + ylab("Frequency") + xlim(50,300)

vari <- COURAGE_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_d)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label_d) + ylab("Frequency") + xlim(20,200)
```

#### Harmonisation algorithm

To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `if dh = 1 and sh = NA into 1`
* `if dh = NA and sh = 1 into 1`
* `other into 999`

**R script**

```{r harmo Spain w1}
COURAGE_ah$dh <- recode(COURAGE_ah$dbp, "0:89.999=0; 90:900=1; 999=NA")
COURAGE_ah$sh <- recode(COURAGE_ah$sbp, "0:139.999=0; 140:900=1; 998=NA; 999=NA")


COURAGE_ds_spain_w1 <- tibble(id=COURAGE_ah$id)
COURAGE_ds_spain_w1$ah <- NA
COURAGE_ds_spain_w1$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==0)] <- 0
COURAGE_ds_spain_w1$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==0)] <- 1
COURAGE_ds_spain_w1$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_spain_w1$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_spain_w1$ah[which(is.na(COURAGE_ah$dh) & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_spain_w1$ah[which(COURAGE_ah$dh==1 & is.na(COURAGE_ah$sh))] <- 1
COURAGE_ds_spain_w1$ah <- car::recode(COURAGE_ds_spain_w1$ah, "NA=999")
COURAGE_ds_spain_w1$ah <- labelled(COURAGE_ds_spain_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript Spain w1, echo=F}
vari <- COURAGE_ds_spain_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(COURAGE_ds_spain_w1, aes(x=factor(ah))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation Spain w1, echo=F}
```


```{r importOPAL Spain w1, echo=FALSE}

label(COURAGE_ds_spain_w1$ah) <- ds_label_all
ah <- COURAGE_ds_spain_w1
save(ah, file = "../RData/spain_w1/ah.RData")
```

### Wave 2

#### Study-specific variable description

The study-specific variables elected to be harmonised is:

* Name: `sbp`
* Label: `Systolic blood pressure of participant at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

* Name: `dbp`
* Label: `Diastolic blood pressure of participantat at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

```{r assign Spain w2, echo=F}
load("../RData/spain_w2/dbp.RData")
COURAGE_ah <- dbp
load("../RData/spain_w2/sbp.RData")
COURAGE_ah$sbp <- sbp$sbp
```

```{r local Spain w2, echo=F}
vari <- COURAGE_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_s)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(x=(sbp))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_s) + ylab("Frequency") + xlim(50,300)

vari <- COURAGE_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_d)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label_d) + ylab("Frequency") + xlim(20,200)
```

#### Harmonisation algorithm

To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `if dh = 993 and sh = 993 into 993`
* `if dh = 997 and sh = 997 into 997`
* `if dh = 999 and sh = 999 into 999`
* `other into 999`

**R script**

```{r harmo Spain w2}
COURAGE_ah$dh <- recode(COURAGE_ah$dbp, "0:89.999=0; 90:900=1")
COURAGE_ah$sh <- recode(COURAGE_ah$sbp, "0:139.999=0; 140:900=1")


COURAGE_ds_spain_w2 <- tibble(id=COURAGE_ah$id)
COURAGE_ds_spain_w2$ah <- NA
COURAGE_ds_spain_w2$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==0)] <- 0
COURAGE_ds_spain_w2$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==0)] <- 1
COURAGE_ds_spain_w2$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_spain_w2$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_spain_w2$ah[which(COURAGE_ah$dh==993 & COURAGE_ah$sh==993)] <- 993
COURAGE_ds_spain_w2$ah[which(COURAGE_ah$dh==997 & COURAGE_ah$sh==997)] <- 997
COURAGE_ds_spain_w2$ah[which(COURAGE_ah$dh==999 & COURAGE_ah$sh==999)] <- 999
COURAGE_ds_spain_w2$ah <- labelled(COURAGE_ds_spain_w2$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript Spain w2, echo=F}
vari <- COURAGE_ds_spain_w2$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(COURAGE_ds_spain_w2, aes(x=factor(ah))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation Spain w2, echo=F}
```




```{r importOPAL Spain w2, echo=FALSE}
#opal.assign.data(o, 'COURAGE_ds_spain_w2', COURAGE_ds_spain_w2)
#opal.symbol_import(o,'COURAGE_ds_spain_w2', project='_Athlos_Harm_Dataset')
ah <- COURAGE_ds_spain_w2
save(ah, file = "../RData/spain_w2/ah.RData")

label(COURAGE_ds_spain_w2$ah) <- ds_label_all
ah <- COURAGE_ds_spain_w2
save(ah, file = "../RData/spain_w2/ah.RData")
```



## Poland 

### Wave 1 

#### Study-specific variable description

The study-specific variables elected to be harmonised is:

* Name: `sbp`
* Label: `Systolic blood pressure of participant at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

* Name: `dbp`
* Label: `Diastolic blood pressure of participantat at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

```{r assign Poland w1, echo=F}
load("../RData/pol_w1/dbp.RData")
COURAGE_ah <- dbp
load("../RData/pol_w1/sbp.RData")
COURAGE_ah$sbp <- sbp$sbp
```

```{r local Poland w1, echo=F}
vari <- COURAGE_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_s)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(x=(sbp))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_s) + ylab("Frequency") + xlim(50,300)

vari <- COURAGE_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_d)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label_d) + ylab("Frequency") + xlim(20,200)
```

#### Harmonisation algorithm

To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `if dh = 999 and sh = 999 into 999`

* `other into 999`

**R script**

```{r harmo Poland w1}
COURAGE_ah$dh <- recode(COURAGE_ah$dbp, "0:89.999=0; 90:900=1")
COURAGE_ah$sh <- recode(COURAGE_ah$sbp, "0:139.999=0; 140:900=1")

COURAGE_ds_pol_w1 <- tibble(id=COURAGE_ah$id)
COURAGE_ds_pol_w1$ah <- NA
COURAGE_ds_pol_w1$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==0)] <- 0
COURAGE_ds_pol_w1$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==0)] <- 1
COURAGE_ds_pol_w1$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_pol_w1$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_pol_w1$ah[which(COURAGE_ah$dh==999 & COURAGE_ah$sh==999)] <- 999
COURAGE_ds_pol_w1$ah <- labelled(COURAGE_ds_pol_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript Poland w1, echo=F}
vari <- COURAGE_ds_pol_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(COURAGE_ds_pol_w1, aes(x=factor(ah))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation Poland w1, echo=F}
```



```{r importOPAL Poland w1, echo=FALSE}
#opal.assign.data(o, 'COURAGE_ds_pol_w1', COURAGE_ds_pol_w1)
#opal.symbol_import(o,'COURAGE_ds_pol_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_pol_w1$ah) <- ds_label_all
ah <- COURAGE_ds_pol_w1
save(ah, file = "../RData/pol_w1/ah.RData")
```

### Wave 2

#### Study-specific variable description

The study-specific variables elected to be harmonised is:

* Name: `sbp`
* Label: `Systolic blood pressure of participant at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

* Name: `dbp`
* Label: `Diastolic blood pressure of participantat at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

```{r assign Poland w2, echo=F}
load("../RData/pol_w2/dbp.RData")
COURAGE_ah <- dbp
load("../RData/pol_w2/sbp.RData")
COURAGE_ah$sbp <- sbp$sbp
```

```{r local Poland w2, echo=F}
vari <- COURAGE_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_s)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(x=(sbp))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_s) + ylab("Frequency") + xlim(50,300)

vari <- COURAGE_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_d)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label_d) + ylab("Frequency") + xlim(20,200)
```

#### Harmonisation algorithm

To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `if dh = 999 and sh = 999 into 999`
* `other into 999`

**R script**

```{r harmo Poland w2}
COURAGE_ah$dh <- recode(COURAGE_ah$dbp, "0:89.999=0; 90:900=1")
COURAGE_ah$sh <- recode(COURAGE_ah$sbp, "0:139.999=0; 140:900=1")


COURAGE_ds_pol_w2 <- tibble(id=COURAGE_ah$id)
COURAGE_ds_pol_w2$ah <- NA
COURAGE_ds_pol_w2$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==0)] <- 0
COURAGE_ds_pol_w2$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==0)] <- 1
COURAGE_ds_pol_w2$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_pol_w2$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_pol_w2$ah[which(COURAGE_ah$dh==999 & COURAGE_ah$sh==999)] <- 999
COURAGE_ds_pol_w2$ah <- labelled(COURAGE_ds_pol_w2$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript Poland w2, echo=F}
vari <- COURAGE_ds_pol_w2$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(COURAGE_ds_pol_w2, aes(x=factor(ah))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation Poland w2, echo=F}
```


```{r importOPAL Poland w2, echo=FALSE}

label(COURAGE_ds_pol_w2$ah) <- ds_label_all
ah <- COURAGE_ds_pol_w2
save(ah, file = "../RData/pol_w2/ah.RData")

```



## Finland 

### Wave 1 

#### Study-specific variable description

The study-specific variables elected to be harmonised is:


* Name: `sbp`
* Label: `Systolic blood pressure of participant at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

* Name: `dbp`
* Label: `Diastolic blood pressure of participantat at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not doneNot applicable`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`

```{r assign Finland w1, echo=F}
load("../RData/fin_w1/dbp.RData")
COURAGE_ah <- dbp
load("../RData/fin_w1/sbp.RData")
COURAGE_ah$sbp <- sbp$sbp
```

```{r local Finland w1, echo=F}
vari <- COURAGE_ah$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_s)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(x=(sbp))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_s) + ylab("Frequency") + xlim(50,300)

vari <- COURAGE_ah$dbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label_d)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(COURAGE_ah, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label_d) + ylab("Frequency") + xlim(20,200)
```

#### Harmonisation algorithm

To compute the harmonized variable, first of all, it has to be created two categorical variables as follows:

* dh:
  + `dbp >= 90 into 1`
  + `dbp < 90 into 0`

* sh:
  + `sbp >= 140 into 1`
  + `sbp < 140 into 0`

After that, to compute the harmonized variable from this two variables has to be recoded as follows:

* `if dh = 0 and sh = 0 into 0`
* `if dh = 1 and sh = 0 into 1`
* `if dh = 0 and sh = 1 into 1`
* `if dh = 1 and sh = 1 into 1`
* `if dh = 999 and sh = 999 into 999`
* `other into 999`

**R script**
```{r harmo Finland w1}
COURAGE_ah$dh <- recode(COURAGE_ah$dbp, "0:89.999=0; 90:900=1")
COURAGE_ah$sh <- recode(COURAGE_ah$sbp, "0:139.999=0; 140:900=1")

COURAGE_ds_fin_w1 <- tibble(id=COURAGE_ah$id)
COURAGE_ds_fin_w1$ah <- NA
COURAGE_ds_fin_w1$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==0)] <- 0
COURAGE_ds_fin_w1$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==0)] <- 1
COURAGE_ds_fin_w1$ah[which(COURAGE_ah$dh==0 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_fin_w1$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_fin_w1$ah[which(COURAGE_ah$dh==999 & COURAGE_ah$sh==1)] <- 1
COURAGE_ds_fin_w1$ah[which(COURAGE_ah$dh==1 & COURAGE_ah$sh==999)] <- 1
COURAGE_ds_fin_w1$ah[which(COURAGE_ah$dh==999 & COURAGE_ah$sh==999)] <- 999
COURAGE_ds_fin_w1$ah <- car::recode(COURAGE_ds_fin_w1$ah, "NA=999")
COURAGE_ds_fin_w1$ah <- labelled(COURAGE_ds_fin_w1$ah, labels = c("no hypertension"=0, "hypertension"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript Finland w1, echo=F}
vari <- COURAGE_ds_fin_w1$ah
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(COURAGE_ds_fin_w1, aes(x=factor(ah))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation Finland w1, echo=F}
```




```{r importOPAL Finland w1, echo=FALSE}
#opal.assign.data(o, 'COURAGE_ds_fin_w1', COURAGE_ds_fin_w1)
#opal.symbol_import(o,'COURAGE_ds_fin_w1', project='_Athlos_Harm_Dataset')

label(COURAGE_ds_fin_w1$ah) <- ds_label_all
ah <- COURAGE_ds_fin_w1
save(ah, file = "../RData/fin_w1/ah.RData")
```







# Quality estimation

No comments.



```{r closeRsession, echo=FALSE}
opal.logout(o)
```
