==> COURAGE/6_Physical_measures/13_heart_rate.Rmd <==
---
title: "Heart Rate"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/heart_rateDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/heart_rateDS.R')
```

# Data process

## Finland

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_pulse`|`q2502a_pulse`|
|-|-|-|
| **Label** |`Pulse(1)`|`Pulse(2)`|
| **Table name** |`Courage_Finland_w1`|`Courage_Finland_w1`|
| **Units** |`bpm`|`bpm`|
| **Description** |  |  |


```{r assign fin_w1, echo=F}
opal.assign.table.tibble(o, 'COURAGE_fin_w1','Courage.Courage_Finland_w1', variables=list('q2501a_pulse','q2502a_pulse'), missings = TRUE)
COURAGE_fin_w1 <- opal.execute(o,'COURAGE_fin_w1')

```

```{r local fin_w1, echo=F}
vbl <- COURAGE_fin_w1$q2501a_pulse
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(COURAGE_fin_w1)[2])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_fin_w1, aes(q2501a_pulse)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") 

vbl <- COURAGE_fin_w1$q2502a_pulse
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(COURAGE_fin_w1)[3])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(COURAGE_fin_w1, aes(q2502a_pulse)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```



#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables:

* `Do the average of the scores to get heart_rate`
* `NA into 999`

**R script:**

```{r harmo fin_w1}

COURAGE_ds_fin_w1 <- tibble(id=COURAGE_fin_w1$id)
COURAGE_ds_fin_w1$heart_rate <- rowMeans(cbind(COURAGE_fin_w1$q2501a_pulse,COURAGE_fin_w1$q2502a_pulse), na.rm = T)
COURAGE_ds_fin_w1$heart_rate <- car::recode(COURAGE_ds_fin_w1$heart_rate, "NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript fin_w1, echo=F}
vbl <- COURAGE_ds_fin_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_fin_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation fin_w1, echo=F}
```






## Poland

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_pulse`|`q2502a_pulse`|
|-|-|-|
| **Label** |`Pulse(1)`|`Pulse(2)`|
| **Table name** |`Courage_Poland_w1`|`Courage_Poland_w1`|
| **Units** |`bpm`|`bpm`|
| **Description** |  |  |


```{r assign pol_w1, echo=F}
opal.assign.table.tibble(o, 'Courage_Poland_w1','Courage.Courage_Poland_w1', variables=list('q2501a_pulse','q2502a_pulse'), missings = TRUE)
Courage_Poland_w1 <- opal.execute(o,'Courage_Poland_w1')

```

```{r local pol_w1, echo=F}
vbl <- Courage_Poland_w1$q2501a_pulse
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(Courage_Poland_w1)[2])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Courage_Poland_w1, aes(q2501a_pulse)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") 

vbl <- Courage_Poland_w1$q2502a_pulse
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(Courage_Poland_w1)[3])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Courage_Poland_w1, aes(q2502a_pulse)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```



#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables:

* `Do the average of the scores to get heart_rate`
* `NA into 999`

**R script:**

```{r harmo pol_w1}

COURAGE_ds_pol_w1 <- tibble(id=Courage_Poland_w1$id)
COURAGE_ds_pol_w1$heart_rate <- rowMeans(cbind(Courage_Poland_w1$q2501a_pulse,Courage_Poland_w1$q2502a_pulse), na.rm = T)
COURAGE_ds_pol_w1$heart_rate <- car::recode(COURAGE_ds_pol_w1$heart_rate, "NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript pol_w1, echo=F}
vbl <- COURAGE_ds_pol_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation pol_w1, echo=F}
```






### Wave 2

#### Study-specific variable description

| **Name** |`q2501a_pulse_s1`|`q2502a_pulse_s1`|
|-|-|-|
| **Label** |`Pulse(1)`|`Pulse(2)`|
| **Table name** |`Courage_Poland_w2`|`Courage_Poland_w2`|
| **Units** |`bpm`|`bpm`|
| **Description** |  |  |


```{r assign pol_w2, echo=F}
opal.assign.table.tibble(o, 'Courage_Poland_w2','Courage.Courage_Poland_w2', variables=list('q2501a_pulse_s1','q2502a_pulse_s1'), missings = TRUE)
Courage_Poland_w2 <- opal.execute(o,'Courage_Poland_w2')

```

```{r local pol_w2, echo=F}
vbl <- Courage_Poland_w2$q2501a_pulse_s1
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(Courage_Poland_w2)[2])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Courage_Poland_w2, aes(q2501a_pulse_s1)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") 

vbl <- Courage_Poland_w2$q2502a_pulse_s1
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(Courage_Poland_w2)[3])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Courage_Poland_w2, aes(q2502a_pulse_s1)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```



#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables:

* `Do the average of the scores to get heart_rate`
* `NA into 999`

**R script:**

```{r harmo pol_w2}

COURAGE_ds_pol_w2 <- tibble(id=Courage_Poland_w2$id)
COURAGE_ds_pol_w2$heart_rate <- rowMeans(cbind(Courage_Poland_w2$q2501a_pulse_s1,Courage_Poland_w2$q2502a_pulse_s1), na.rm = T)
COURAGE_ds_pol_w2$heart_rate <- car::recode(COURAGE_ds_pol_w2$heart_rate, "NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript pol_w2, echo=F}
vbl <- COURAGE_ds_pol_w2$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_pol_w2, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation pol_w2, echo=F}
```






## Spain

### Wave 1

#### Study-specific variable description

| **Name** |`q2501a_pulse`|`q2502a_pulse`|
|-|-|-|
| **Label** |`Pulse(1)`|`Pulse(2)`|
| **Table name** |`Courage_Spain_w1`|`Courage_Spain_w1`|
| **Units** |`bpm`|`bpm`|
| **Description** |  |  |


```{r assign spain_w1, echo=F}
opal.assign.table.tibble(o, 'Courage_Spain_w1','Courage.Courage_Spain_w1', variables=list('q2501a_pulse','q2502a_pulse'), missings = TRUE)
Courage_Spain_w1 <- opal.execute(o,'Courage_Spain_w1')

```

```{r local spain_w1, echo=F}
vbl <- Courage_Spain_w1$q2501a_pulse
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(Courage_Spain_w1)[2])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Courage_Spain_w1, aes(q2501a_pulse)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vbl)) + ylab("Frequency") 

vbl <- Courage_Spain_w1$q2502a_pulse
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(Courage_Spain_w1)[3])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Courage_Spain_w1, aes(q2502a_pulse)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```



#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables:

* `Do the average of the scores to get heart_rate`
* `NA into 999`

**R script:**

```{r harmo spain_w1}

COURAGE_ds_spain_w1 <- tibble(id=Courage_Spain_w1$id)
COURAGE_ds_spain_w1$heart_rate <- rowMeans(cbind(Courage_Spain_w1$q2501a_pulse,Courage_Spain_w1$q2502a_pulse), na.rm = T)
COURAGE_ds_spain_w1$heart_rate <- car::recode(COURAGE_ds_spain_w1$heart_rate, "NA=999")

```

#### Statistical description of the new harmonised variable
```{r descript spain_w1, echo=F}
vbl <- COURAGE_ds_spain_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_spain_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation spain_w1, echo=F}
```




### Wave 2

#### Study-specific variable description

| **Name** |`q2501a_pulse_s1`|`q2502a_pulse_s1`|
|-|-|-|
| **Label** |`Pulse(1)`|`Pulse(2)`|
| **Table name** |`Courage_Spain_w2_1`|`Courage_Spain_w2_1`|
| **Units** |`bpm`|`bpm`|
| **Missings** |`997 = Disable to measure`<br/>`998 = Refusal`|  |
| **Description** |  |  |


```{r assign spain_w2, echo=F}
opal.assign.table.tibble(o, 'Courage_Spain_w2_1','Courage.Courage_Spain_w2_1', variables=list('q2501a_pulse_s1','q2502a_pulse_s1'), missings = TRUE)
Courage_Spain_w2_1 <- opal.execute(o,'Courage_Spain_w2_1')

```

```{r local spain_w2, echo=F}
vbl <- Courage_Spain_w2_1$q2501a_pulse_s1
vbl_miss <- c(997,998)
kable(Continuous_summary(vbl, missing_values = vbl_miss)[3], caption = names(Courage_Spain_w2_1)[2])
pander(Continuous_summary(vbl, missing_values = vbl_miss)$summary, caption = "Summary")
ggplot(Courage_Spain_w2_1, aes(q2501a_pulse_s1)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% vbl_miss)) + xlab(label(vbl)) + ylab("Frequency") 

vbl <- Courage_Spain_w2_1$q2502a_pulse_s1
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = names(Courage_Spain_w2_1)[3])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Courage_Spain_w2_1, aes(q2502a_pulse_s1)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```



#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variables:

* `Do the average of the scores to get heart_rate`
* `NA into 999`

**R script:**

```{r harmo spain_w2}

COURAGE_ds_spain_w2 <- tibble(id=Courage_Spain_w2_1$id)
COURAGE_ds_spain_w2$heart_rate <- rowMeans(cbind(Courage_Spain_w2_1$q2501a_pulse_s1,Courage_Spain_w2_1$q2502a_pulse_s1), na.rm = T)
COURAGE_ds_spain_w2$heart_rate <- car::recode(COURAGE_ds_spain_w2$heart_rate, "NA=999")
COURAGE_ds_spain_w2$heart_rate[which(Courage_Spain_w2_1$q2501a_pulse_s1==998)] <- car::recode(Courage_Spain_w2_1$q2502a_pulse_s1[which(Courage_Spain_w2_1$q2501a_pulse_s1==998)],"NA=997")
COURAGE_ds_spain_w2$heart_rate[which(Courage_Spain_w2_1$q2501a_pulse_s1==997)] <- car::recode(Courage_Spain_w2_1$q2502a_pulse_s1[which(Courage_Spain_w2_1$q2501a_pulse_s1==997)],"NA=993")

```

#### Statistical description of the new harmonised variable
```{r descript spain_w2, echo=F}
vbl <- COURAGE_ds_spain_w2$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(COURAGE_ds_spain_w2, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation spain_w2, echo=F}
```





## Summary of descriptive statistics of the harmonised variable accross populations and waves

```{r helpfunctions, echo=F}

labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with continuous data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cont_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for continuous data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of column with summary table categories
  t.summ <- summary(m.hds[[1]][2])[1:6]
  # Adding of missing/no-missing values categories
  t.hds <- c(substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1),labels(Continuous_summary(l.hds[[1]][[2]],missing_values = miss_values_vector)$values_table)[[2]])
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    # First, summary values
    t.summ <- summary(m.hds[[i]][2])[1:6]
    # Next, missing/no-missing values
    t.hds <- cbind(t.hds,c(as.numeric(substr(t.summ,regexpr(":", t.summ, fixed=T)+1,nchar(t.summ))),as.numeric(Continuous_summary(l.hds[[i]][[2]],missing_values = miss_values_vector)$values_table[1,])))
  }
  # Add sample size for each wave/population
  t.hds <- rbind(t.hds,c("sample size", sapply(l.hds,function(wave) length(wave[[1]]))))
  # Add wave/population names
  dimnames(t.hds)[[2]] <- c(dimnames(summary(m.hds[[1]][2]))[[2]],lnames)
  return(t.hds)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


courage.cw <- list(fin_w1 = c("fin_w1","Finland w1"), pol_w1 = c("pol_w1","Poland w1"), pol_w2 = c("pol_w2","Poland w2"), spain_w1 = c("spain_w1","Spain w1"), spain_w2 = c("spain_w2","Spain w2"))

# Consider only harmonised waves
l.hds <- lapply(courage.cw, function(wname) if(exists(paste0("COURAGE_ds_",wname[1]))){wname = list(get(paste0("COURAGE_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing summaries
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```





# Quality estimation



<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
